add cast fix intermittent compilation error from junit upgrade
git-svn-id: https://svn.apache.org/repos/asf/lucene/solr/branches/solr@923401 13f79535-47bb-0310-9956-ffa450edef68

+++ contrib/dataimporthandler/src/test/java/org/apache/solr/handler/dataimport/TestSqlEntityProcessor.java
@@ -122,7 +122,7 @@ public class TestSqlEntityProcessor {
         break;
       count++;
     }
-    Assert.assertEquals(2, local.get());
+    Assert.assertEquals(2, (int) local.get());
     Assert.assertEquals(4, count);
   }
